pyproject.toml
[project]
name = "agentic-rag-manufacturing-investment"
version = "0.1.0"
description = "Agentic RAG workflow for evaluating manufacturing AI startup investments"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.1.1",
    "langchain>=1.0.2",
    "langchain-chroma>=1.0.0",
    "langchain-huggingface>=1.0.1",
    "langchain-openai>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.1",
    "pydantic>=2.11.3",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "tavily-python>=0.7.12",
    "typing-extensions>=4.13.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
